feat(util): reorganize `scripts`-related functions
